Add benchmark for ECEF-to-NED/ENU rotation construction
Measured at 0.10.5 on aarch64-darwin, M4 pro: `ecef_to_ned_at` about
957 ns and `ecef_to_enu_at` about 930 ns per call.
